DC 32 Game Buttons D-Pad and A B
Files added for fusion archive files, stl, and print file for both on orca slicer for 15x copies. The print file was done on the fly, so use with caution. Buttons had to be widened out a little. Superglue should hold these in place just fine if you're not comfortable with the friction fit.